This is code cleanup that breaks things up into smaller modules with the intent of adding additional features like updating binding syntax.
