+++ .travis.yml
@@ -1,0 +1,5 @@
+language: node_js
+node_js: 0.12
+before_install:
+  - "export DISPLAY=:99.0"
+  - "sh -e /etc/init.d/xvfb start"
